agents/requirements.txt
# CrewAI Framework with Tools
crewai[tools]==1.3.0

# Data Validation
pydantic>=2.0.0

# YAML Configuration Parsing
pyyaml>=6.0.0
